+++ CHANGES.rst
@@ -5,10 +5,11 @@ Version 8.4.1
 
 Unreleased
 
+-   ``get_parameter_source()`` is available during eager callbacks and type
+    conversion again. :issue:`3458` :issue:`3484`
 -   Zsh completion scripts parse correctly on Windows. :issue:`3277`
 -   Shell completion of `Choice` `Enum` values produces a valid completion
     result. :issue:`3015`
-
 
 Version 8.4.0
 -------------